move: get_sql_start_date from export to migration util
