+++ corehq/apps/auditcare/utils/migration.py
@@ -6,13 +6,28 @@ from django.core.cache import cache
 
 from dimagi.utils.dates import force_to_datetime
 
-from corehq.apps.auditcare.models import AuditcareMigrationMeta
-from corehq.apps.auditcare.utils.export import get_sql_start_date
+from corehq.apps.auditcare.models import AuditcareMigrationMeta, NavigationEventAudit
 
 CUTOFF_TIME = datetime(2012, 12, 31)
 CACHE_TTL = 14 * 24 * 60 * 60  # 14 days
 
 logger = logging.getLogger(__name__)
+
+
+def get_sql_start_date():
+    """Get the date of the first SQL auditcare record
+
+    HACK this uses `NavigationEventAudit` since that model is likely to
+    have the record with the earliest timestamp.
+
+    NOTE this function assumes no SQL data has been archived, and that
+    all auditcare data in Couch will be obsolete and/or archived before
+    SQL data. It should be removed when the data in Couch is no longer
+    relevant.
+    """
+    manager = NavigationEventAudit.objects
+    row = manager.order_by("event_date").values("event_date")[:1].first()
+    return row["event_date"] if row else datetime.utcnow()
 
 
 class AuditCareMigrationUtil():
